actividad4/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
mysql-connector-python
Flask-WTF
